Refactor code for CircularBuffer and KillerSudokuHelper, remove SudokuSolver file
Refactored the CircularBuffer code to extend the ArrayDeque class to streamline various operations. Also, the standalone SudokuSolver file was removed, and its relevant parts were integrated as a record within the KillerSudokuHelper class for more efficient architecture.
